+++ web.js
@@ -1,18 +1,18 @@
-'use strict';
-
 module.exports = {
 	extends: [
-		'kink'
+		'kink',
 	],
 	env: {
 		browser: true,
 		jquery: true,
 		mocha: true,
 		mongo: false,
-		node: false
+		node: false,
+	},
+	parserOptions: {
+		sourceType: 'script',
 	},
 	rules: {
-
 		// https://eslint.org/docs/rules/id-blacklist
 		// Added as a step to avoid scripts triggering ad-blockers.
 		'id-blacklist': [
@@ -21,15 +21,15 @@ module.exports = {
 			'ads',
 			'banner',
 			'banners',
-			'social'
+			'social',
 		],
 
 		// Whitelisting done for 'alternative promise chains'.
 		'promise/catch-or-return': [
 			'warn',
 			{
-				terminationMethod: ['catch', 'done']
-			}
-		]
-	}
+				terminationMethod: ['catch', 'done'],
+			},
+		],
+	},
 };