+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG.md
 
+## v1.10.0 - 2019.03.07
+
+- Set `parserOptions` `sourceType` to `module` in `index` (default), `script` in `vue` and `web`
+- Updated `no-multiple-empty-lines` rule to include `maxBOF` and `maxEOF` options
+- Updated `comma-dangle` to require `always-multiline` for arrays and objects 
+- Updated `eslint` to `5.15.x`, `eslint-plugin-jsdoc` to `4.1.x`, `eslint-plugin-mocha` to `5.3.x`, `eslint-plugin-vue` to `5.2.x`
+
 ## v1.9.0 - 2019.01.18
 
 - Updated `eslint-plugin-jsdoc` to v4, `eslint-plugin-node` to v8